adds delete authorization and update authorization

+++ App/controllers/ListingController.php
@@ -2,7 +2,9 @@
 
 namespace App\Controllers;
 
+use Framework\Authorization;
 use Framework\Database;
+use Framework\Session;
 use Framework\Validation;
 
 class ListingController
@@ -20,11 +22,9 @@ class ListingController
      *
      * @return void
      */
-
     public function index()
     {
-
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
         loadView('listings/index', [
             'listings' => $listings,
@@ -36,7 +36,6 @@ class ListingController
      *
      * @return void
      */
-
     public function create()
     {
         loadView('listings/create');
@@ -50,7 +49,6 @@ class ListingController
      */
     public function show($params)
     {
-
         $id = $params['id'] ?? '';
 
         $params = [
@@ -75,7 +73,6 @@ class ListingController
      */
     public function store()
     {
-
         $allowedFields = [
             'title',
             'description',
@@ -93,7 +90,7 @@ class ListingController
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -141,17 +138,17 @@ class ListingController
                 if ($value === '') {
                     $value = null;
                 } else {
-
                     $values[] = ':' . $field;
                 }
-
             }
 
             $values = implode(', ', $values);
 
-            $query = "INSERT INTO listings ({$fields}) VALUES ({$values})";
+            $query = "INSERT INTO listings ({$fields}, created_at) VALUES ({$values}, NOW())";
 
             $this->db->query($query, $newListingData);
+
+            Session::setFlashMessage('success_message', 'Listing created successfully');
 
             redirect('listings');
         }
@@ -172,15 +169,23 @@ class ListingController
 
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+        // Check if listing exists
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
         }
 
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to delete this listing');
+
+            return redirect('/test-project/workopia/public/listings/' . $listing->id);
+        }
+
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
         // Set flash message
-        $_SESSION['success_message'] = 'Listing deleted successfully';
+        Session::setFlashMessage('success_message', 'Listing deleted successfully');
 
         redirect('/test-project/workopia/public/listings');
     }
@@ -206,6 +211,11 @@ class ListingController
             return;
         }
 
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to delete this listing');
+            return redirect('/test-project/workopia/public/listings/' . $listing->id);
+        }
+
         loadView('listings/edit', [
             'listing' => $listing,
         ]);
@@ -219,16 +229,21 @@ class ListingController
      */
     public function update($params)
     {
-
         $id = $params['id'] ?? '';
         $params = [
             'id' => $id,
         ];
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
-// Check if listing exists
+        // Check if listing exists
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+            return redirect('/test-project/workopia/public/listings/' . $listing->id);
         }
 
         $allowedFields = [
@@ -283,7 +298,7 @@ class ListingController
             // Submit to database
             $updateFields = [];
 
-            foreach(array_keys($updatedValues) as $field) {
+            foreach (array_keys($updatedValues) as $field) {
                 $updateFields[] = "{$field} = :{$field}";
             }
 
@@ -295,12 +310,31 @@ class ListingController
 
             $this->db->query($updateQuery, $updatedValues);
 
-            $_SESSION['success_message'] = 'Listing updated successfully';
+            Session::setFlashMessage('success_message', 'Listing updated successfully');
 
             redirect('/test-project/workopia/public/listings/' . $id);
-
-
         }
+    }
+
+    /**
+     * Search listings by keywords/location
+     *
+     * @return void
+     */
+    public function search() {
+        $keywords = isset($_GET['keywords']) ? trim($_GET['keywords']) : '';
+        $location = isset($_GET['location']) ? trim($_GET['location']) : '';
+
+        $query = "SELECT * FROM listings WHERE (title LIKE :keywords OR description LIKE :keywords OR tags LIKE :keywords OR company LIKE :keywords) AND (city LIKE :location OR state LIKE :location)";
+
+        $params = [
+            'keywords' => "%{$keywords}%",
+            'location' => "%{$location}%"
+        ];
+
+        $listings = $this->db->query($query, $params)->fetchAll();
+
+        inspectAndDie($listings);
 
     }
 }
